Allow looping over over only certain 'type's of pages ...
I want to put files in content/camaro and content/posts that aren't
posts and shouldn't be in these lists. With this change, to be included
in the list, add "type: camaro" or "type: post" to the hugo header of
the content page. Make it something else if you don't want to include
it.