Smaller line height, plus fix for code block bug ...?
The font for lines in code blocks on my iPhone 11 that cause the
div/span/whatever to overflow was slightly larger.

Trying random shit until it git fixed ... turns out this did it? Found
these options on this blog post for what seems to be an unrelated
problem.

https://kilianvalkhof.com/2022/css-html/your-css-reset-needs-text-size-adjust-probably/

The text-size-adjust options did it.

I also wanted a smaller line-height. That's unrelated :p